+++ scripts/repl.ts
@@ -15,14 +15,14 @@ import { DateTime } from 'luxon';
 
 import { syncCoSouls } from '../api/hasura/cron/syncCoSouls';
 import { updateProfileNFTs } from '../api/nfts/alchemy';
-import { adminClient as client } from '../api-lib/gql/adminClient';
-import { genPgives } from '../api-lib/pgives';
-import { syncPoapDataForCoLinksUsers } from '../api-lib/poap/poap-api';
 import {
   sendEpochEndedEmail,
   sendEpochEndingSoonEmail,
   sendEpochStartedEmail,
-} from '../api-lib/postmark';
+} from '../api-lib/email/postmark';
+import { adminClient as client } from '../api-lib/gql/adminClient';
+import { genPgives } from '../api-lib/pgives';
+import { syncPoapDataForCoLinksUsers } from '../api-lib/poap/poap-api';
 import {
   getOnChainPGIVE,
   getTokenId,
